requirements.txt
Flask
requests
python-dotenv
gspread
oauth2client
sendgrid
streamlit
